+++ src/callModel.ts
@@ -43,7 +43,7 @@ function removePromptsFromCompletions (completions: ModelCompletion[], prompt: s
   });
 }
 
-export async function getModelResponse(prompt: string, suffix: string, infill: boolean, scorePromptOnly: boolean = false): Promise<ModelResponse | null> {
+export async function getModelResponse(prompt: string, suffix: string | null, scorePromptOnly: boolean = false): Promise<ModelResponse | null> {
   const workbenchConfig = vscode.workspace.getConfiguration('worldspider');
   const model = workbenchConfig.get('generation.model');
   const numCompletions = scorePromptOnly ? 1 : workbenchConfig.get('generation.numCompletions');
@@ -57,6 +57,8 @@ export async function getModelResponse(prompt: string, suffix: string, infill: b
   const log = workbenchConfig.get('log');
   const apiKey = workbenchConfig.get('apiKey');
 
+  const chat = model === 'gpt-3.5-turbo' || model === 'gpt-4';
+
   if(!apiKey) {
     vscode.window.showErrorMessage("No API key set. Please set the 'worldspider.apiKey' setting.");
     return null;
@@ -67,25 +69,43 @@ export async function getModelResponse(prompt: string, suffix: string, infill: b
   });
   const openai = new OpenAIApi(configuration);
 
-  try {
-    const response = await openai.createCompletion({
-      model: model,
-      prompt: prompt,
-      suffix: infill ? suffix : null,
-      max_tokens: maxTokens,
-      top_p: topP,
-      frequency_penalty: frequencyPenalty,
-      presence_penalty: presencePenalty,
-      logprobs: logprobs,
-      echo: echo,
-      n: numCompletions,
-      temperature: temperature,
+  const request = {
+    model: model,
+    // prompt: prompt,
+    max_tokens: maxTokens,
+    top_p: topP,
+    frequency_penalty: frequencyPenalty,
+    presence_penalty: presencePenalty,
+    n: numCompletions,
+    temperature: temperature,
+    ...(!chat && {logprobs: logprobs}),
+    ...(!chat && {echo: echo}),
+    ...(!chat && {suffix: suffix}),
+    ...(chat && { messages: [{ role: "assistant", content: prompt }] }),
+    ...(!chat && { prompt }),
+  };
+
+  try {      
+    const response = chat ? await openai.createChatCompletion(request) : await openai.createCompletion(request);
+
+    // if chat, change the format of the response from choices[i].message.content to choices[i].text
+
+    const choices = response.data.choices.map((choice: any) => {
+      if(chat) {
+        return {
+          ...choice,
+          text: " " + choice.message.content, // TODO remove this space and find less hacky way to fix the issue
+        };
+      } else {
+        return choice;
+      }
     });
+    
     const responseObject = {
       ...response.data,
-      choices: echo ? removePromptsFromCompletions(response.data.choices, prompt) : response.data.choices,
+      choices: echo ? removePromptsFromCompletions(choices, prompt) : choices,
       prompt: prompt,
-      suffix: infill ? suffix : null,
+      suffix: suffix,
       echo: echo,
     };
     if(log) {
@@ -96,6 +116,62 @@ export async function getModelResponse(prompt: string, suffix: string, infill: b
     console.log(error);
     let errorMessage = (error as Error).message;
     vscode.window.showErrorMessage(`model call failed: ${errorMessage}`);
+    return null;
+  }
+}
+
+
+export async function modelOperation(prefix:string, suffix: string, selectedText: string ) {
+  const workbenchConfig = vscode.workspace.getConfiguration('worldspider');
+
+  const apiKey = workbenchConfig.get('apiKey');
+  const instruction = workbenchConfig.get('generation.customInstruction');
+  const model = workbenchConfig.get('generation.customOpModel');
+
+  const messages = [
+    { role: "system", content: instruction },
+    { role: "user", content: selectedText }
+  ];
+
+  const request = {
+    model: model,
+    max_tokens: 200,
+    n: 3,
+    temperature: 1.2,
+    messages: messages,
+  };
+
+  const configuration = new Configuration({
+    apiKey: apiKey,
+  });
+  const openai = new OpenAIApi(configuration);
+
+  // console.log(request);
+
+  try {
+    const response = await openai.createChatCompletion(request);
+
+    const choices = response.data.choices.map((choice: any) => {
+      return {
+        ...choice,
+        text: choice.message.content,
+      };
+    });
+
+    console.log(choices);
+
+    const responseObject = {
+      ...response.data,
+      choices: choices,
+      prefix: prefix,
+      suffix: suffix,
+    };
+
+    return responseObject;
+  }
+  catch (error) {
+    console.log(error);
+    vscode.window.showErrorMessage("model call failed");
     return null;
   }
 }